requirements-api.txt
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# OpenAI and LangChain
openai==1.3.5
langchain==0.0.335
langchain-openai==0.0.2
langchain-community==0.0.2

# Vector store
chromadb==0.4.18

# Utilities
python-dotenv==1.0.0

# Existing dependencies from your project
requests==2.31.0
beautifulsoup4==4.12.2
markdownify==0.11.6
tqdm==4.66.1